requirements.txt
asgiref==3.8.1
Django==5.0.4
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
